Message id is only 28 bits, no need for long
This reverts commit e99ecad29da7702858b189c87c0bc26b643ca894.

Conflicts:

	src/freenet/node/NewPacketFormat.java

+++ src/freenet/node/MessageFragment.java
@@ -4,13 +4,13 @@ class MessageFragment {
 	boolean shortMessage;
 	boolean isFragmented;
 	boolean firstFragment;
-	long messageID;
+	int messageID;
 	int fragmentLength;
 	int messageLength;
 	int fragmentOffset;
 	byte[] fragmentData;
 
-	public MessageFragment(boolean shortMessage, boolean isFragmented, boolean firstFragment, long messageID,
+	public MessageFragment(boolean shortMessage, boolean isFragmented, boolean firstFragment, int messageID,
 	                int fragmentLength, int messageLength, int fragmentOffset, byte[] fragmentData) {
 		this.shortMessage = shortMessage;
 		this.isFragmented = isFragmented;
